refactor: update sdk interface for clarity
