+++ crates/sdk/src/prover/evm.rs
@@ -13,7 +13,7 @@ use openvm_verify_stark_host::VmStarkProof;
 use crate::prover::Halo2Prover;
 use crate::{
     prover::{
-        vm::types::VmProvingKey, AggProver, DeferralPathProver, InternalLayerMetadata, RootProver,
+        vm::types::VmProvingKey, AggProver, DeferralAggProver, InternalLayerMetadata, RootProver,
         StarkProver,
     },
     DeferralInput, StdIn, SC,
@@ -48,7 +48,7 @@ where
         app_vm_pk: &VmProvingKey<VB::VmConfig>,
         app_exe: Arc<VmExe<Val<SC>>>,
         agg_prover: Arc<AggProver>,
-        def_prover: Option<Arc<DeferralPathProver>>,
+        def_prover: Option<Arc<DeferralAggProver>>,
         root_prover: Arc<RootProver>,
         #[cfg(feature = "evm-prove")] halo2_prover: Option<Halo2Prover>,
     ) -> Result<Self> {
