+++ docs/vocs/docs/pages/book/acceleration-using-extensions/deferral.mdx
@@ -55,9 +55,9 @@ The deferral extension cannot be configured via `openvm.toml` alone — it requi
 
 To enable deferrals in your VM configuration:
 
-1. Build a `DeferralProver` with the appropriate single-circuit prover for your use case.
-2. Create the extension with `DeferralProver::make_extension`, passing the deferral function(s).
+1. Build a `MultiDeferralCircuitProver` with the appropriate single-circuit prover for your use case.
+2. Create the extension with `MultiDeferralCircuitProver::make_extension`, passing the deferral function(s).
 3. Set `SdkVmConfig.deferral = Some(deferral_ext)`.
-4. Build the SDK with `Sdk::builder().deferral_prover(deferral_prover)`.
+4. Build the SDK with `Sdk::builder().multi_deferral_circuit_prover(multi_deferral_circuit_prover)`.
 
 See the [Verify STARK guest library](/book/guest-libraries/verify-stark#using-deferralextension) for a complete example of wiring up the deferral extension with the STARK verification circuit.
